draft for regions support
