+++ inseeds/components/base/__init__.py
@@ -3,6 +3,6 @@ from .output import Output
 from .entity import Entity
 from .individual import Individual
 from .cell import Cell
+from .region import Region, Country, WorldRegion
 from .world import World
-
 from .component import Component
