+++ inseeds/components/lpjml/__init__.py
@@ -1,3 +1,4 @@
 from pycopanlpjml import Cell
+from pycopanlpjml import Region, Country, WorldRegion
 from pycopanlpjml import World
 from .component import Component
